refactor: ChoiceContent 응답 모델을 factory가 아닌 content 클래스 내에서 생성한다
