+++ src/main/java/life/offonoff/ab/domain/topic/choice/content/ChoiceContent.java
@@ -2,6 +2,7 @@ package life.offonoff.ab.domain.topic.choice.content;
 
 import jakarta.persistence.*;
 import life.offonoff.ab.domain.BaseEntity;
+import life.offonoff.ab.web.response.topic.choice.content.ChoiceContentResponse;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -15,4 +16,6 @@ public abstract class ChoiceContent extends BaseEntity {
 
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
+
+    public abstract ChoiceContentResponse toResponse();
 }
