+++ src/main/java/life/offonoff/ab/domain/topic/choice/content/ImageTextChoiceContent.java
@@ -1,6 +1,10 @@
 package life.offonoff.ab.domain.topic.choice.content;
 
-import jakarta.persistence.*;
+import jakarta.persistence.Column;
+import jakarta.persistence.DiscriminatorValue;
+import jakarta.persistence.Entity;
+import life.offonoff.ab.web.response.topic.choice.content.ChoiceContentResponse;
+import life.offonoff.ab.web.response.topic.choice.content.ChoiceContentResponse.ImageTextChoiceContentResponse;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -17,5 +21,10 @@ public class ImageTextChoiceContent extends ChoiceContent {
     public ImageTextChoiceContent(String imageUrl, String text) {
         this.imageUrl = imageUrl;
         this.text = text;
+    }
+
+    @Override
+    public ChoiceContentResponse toResponse() {
+        return new ImageTextChoiceContentResponse(imageUrl, text);
     }
 }
