+++ src/main/java/life/offonoff/ab/web/response/topic/choice/ChoiceResponse.java
@@ -2,16 +2,15 @@ package life.offonoff.ab.web.response.topic.choice;
 
 import life.offonoff.ab.domain.topic.choice.Choice;
 import life.offonoff.ab.domain.topic.choice.ChoiceOption;
-import life.offonoff.ab.web.response.topic.choice.content.ChoiceContentResponseFactory;
-import life.offonoff.ab.web.response.topic.choice.content.ChoiceContentResponseFactory.ChoiceContentResponse;
+import life.offonoff.ab.web.response.topic.choice.content.ChoiceContentResponse;
 import lombok.Getter;
 
 @Getter
 public class ChoiceResponse {
 
-    private Long choiceId;
-    private ChoiceContentResponse content;
-    private ChoiceOption choiceOption;
+    private final Long choiceId;
+    private final ChoiceContentResponse content;
+    private final ChoiceOption choiceOption;
 
     public ChoiceResponse(Long choiceId, ChoiceContentResponse content, ChoiceOption choiceOption) {
         this.choiceId = choiceId;
@@ -22,7 +21,7 @@ public class ChoiceResponse {
     public static ChoiceResponse from(Choice choice) {
         return new ChoiceResponse(
                 choice.getId(),
-                ChoiceContentResponseFactory.create(choice.getContent()),
+                choice.generateContentResponse(),
                 choice.getChoiceOption());
     }
 }
