+++ src/main/java/life/offonoff/ab/web/response/topic/choice/ChoiceResponseWithCount.java
@@ -1,17 +1,16 @@
 package life.offonoff.ab.web.response.topic.choice;
 
 import life.offonoff.ab.domain.topic.choice.Choice;
-import life.offonoff.ab.web.response.topic.choice.content.ChoiceContentResponseFactory;
 import lombok.Getter;
 
 @Getter
 public class ChoiceResponseWithCount extends ChoiceResponse {
 
-    private int voteCount;
+    private final int voteCount;
 
     public ChoiceResponseWithCount(Choice choice) {
         super(choice.getId(),
-              ChoiceContentResponseFactory.create(choice.getContent()),
+              choice.generateContentResponse(),
               choice.getChoiceOption());
 
         this.voteCount = choice.getVoteCount();
